small refactor to polynomial commitment tests
